fix(geometry): drain the boolean processor's failure log into the router (#3821)
`BooleanClippingProcessor::take_failures` was called from tests only. Every
record it made — an unsupported operand, an `EmptyOperand` cutter, an unknown
operator — went into a `RefCell` no production caller ever read, so a model
whose booleans had all degraded still reported a clean load. Absence was
indistinguishable from success.

The base-operand arm was worse: `_ => Ok(Mesh::new())` in
`process_operand_with_depth` meshed an unsupported operand as an empty mesh and
recorded nothing at all. Returning `Err` there is not the fix — for a SECOND
operand that arm means "unsupported cutter, host renders uncut", and an error
would delete the host's item too — so the arm keeps its empty mesh and now
records `BoolFailureReason::UnsupportedOperand(<ifc type>)`.

- `GeometryProcessor::take_bool_failures` (defaulted to none) is the drain hook;
  `BooleanClippingProcessor` overrides it.
- `GeometryRouter::drain_processor_failures` sweeps every registered processor's
  log into `csg_failures`, and `take_csg_failures` calls it — so the native
  pipeline and the wasm batch path both pick these up with no second opt-in.
  Bucketed under product id 0, the same "no attribution" bucket
  `PENDING_MAPPED_BOOL_FAILURES` already uses; both pipelines drain per element,
  so the records stay scoped to the right element in the aggregate.
- `CsgSolidProcessor` builds a TRANSIENT boolean processor per
  `IfcCsgSolid.TreeRootExpression` that the router's table cannot reach, so it
  hands its log to that same thread-local escape hatch instead of dropping it.

Behaviour is unchanged: the meshes produced are byte-identical, this only makes
the drops visible.

Module-size ratchet: split rather than raised. `GeometryProcessor` moved to
router/processor.rs (749 -> 729), `process_operand_with_depth` to
boolean/operand.rs (1005 -> 951), the sweep to
router/diagnostics/processor_failures.rs, and `IfcSphere` to processors/sphere.rs
(csg_primitive.rs 422 -> 324, row deleted). router/diagnostics.rs also lost the
second, hand-maintained copy of the `BoolFailureReason` label match, which
nothing but prose kept in step with `BoolFailureReason::label()`.

RED/GREEN, by mutation: dropping the `take_csg_failures` sweep, the record at
the operand arm, or the CsgSolid hand-off each fails a distinct test in
tests/issue_3821_router_drains_bool_failures.rs; the clean-model control stays
green under all three.
